requirements.txt
hydra-core>=1.3.0

# 画像モデルライブラリ (必須)
timm>=0.9.0

# ログ管理 (WandBを使う予定があれば)
wandb

# データ処理・プログレスバー (Kaggleには入っていますが念のため)
pandas
scikit-learn
tqdm